Added optional note field to default header
The note is configurable via the extension settings. It does not show up in the copyright header if it is not set. If it is set, it appears on a new line at the end of the default copyright notice header. For now, this feature is only available on the default header. If there is desire for custom notes on Apache, MIT, etc... headers that can be added in the future.