Add support for non-C-style languages
Languages like Python which use # for comments instead of // can now be supported. I've specified Python files and Shell scripts as options
for this currently, but the process of adding more is hopefully relatively self-explanatory (but instead of adding the lang-id to the configuredLanguages enum,
add it to the appropriate language style enum).